chore(components): rename typography components
