+++ packages/components/src/components/Modal/Modal.tsx
@@ -5,7 +5,7 @@ import styled, { css } from 'styled-components';
 import { useKeyPress } from '@trezor/react-utils';
 import { useTheme } from '../../utils/hooks';
 import { Icon } from '../Icon/Icon';
-import { H1 } from '../typography/Heading';
+import { H1 } from '../typography/Heading/Heading';
 import { variables } from '../../config';
 import { IconType } from '../../support/types';
 import { Progress } from '../loaders/Progress';
