+++ packages/components/src/index.ts
@@ -23,9 +23,9 @@ export * from './components/form/Range/Range';
 export * from './components/form/Switch';
 export * from './components/form/InputStyles';
 export * from './components/Icon/Icon';
-export * from './components/typography/Heading';
-export * from './components/typography/Paragraph';
-export * from './components/typography/Link';
+export * from './components/typography/Heading/Heading';
+export * from './components/typography/Paragraph/Paragraph';
+export * from './components/typography/Link/Link';
 export * from './components/Warning/Warning';
 export * from './components/others/Box';
 export * from './components/others/Card';
